Fix linked list header file
All functions were declared as `inline static` which means
they must be defined in the header file.
