--- kernel/lib/linked_list.cpp
@@ -1,85 +1,0 @@
-/**
- * @file linked_list.cpp
- * @author Keeton Feavel (keetonfeavel@cedarville.edu)
- * @brief Standard linked list library
- * @version 0.1
- * @date 2020-06-17
- *
- * @copyright Copyright Keeton Feavel et al (c) 2020
- *
- */
-
-#include <lib/linked_list.hpp>
-
-// initialize a one element DList
-static inline void dlist_init(DList *dlist) {
-	//printf("%s(%p)\n", __FUNCTION__, dlist);
-    dlist->next = dlist;
-    dlist->prev = dlist;
-}
-
-// insert d2 after d1
-static inline void dlist_insert_after(DList *d1, DList *d2) {
-	//printf("%s(%p, %p)\n", __FUNCTION__, d1, d2);
-    DList *n1 = d1->next;
-    DList *e2 = d2->prev;
-
-    d1->next = d2;
-    d2->prev = d1;
-    e2->next = n1;
-    n1->prev = e2;
-}
-
-// insert d2 before d1
-static inline void dlist_insert_before(DList *d1, DList *d2) {
-	//printf("%s(%p, %p)\n", __FUNCTION__, d1, d2);
-    DList *e1 = d1->prev;
-    DList *e2 = d2->prev;
-
-    e1->next = d2;
-    d2->prev = e1;
-    e2->next = d1;
-    d1->prev = e2;
-}
-
-// remove d from the list
-static inline void dlist_remove(DList *d) {
-	//printf("%s(%p)\n", __FUNCTION__, d);
-    d->prev->next = d->next;
-    d->next->prev = d->prev;
-    d->next = d;
-    d->prev = d;
-}
-
-// push d2 to the front of the d1p list
-static inline void dlist_push(DList **d1p, DList *d2) {
-	//printf("%s(%p, %p)\n", __FUNCTION__, d1p, d2);
-    if (*d1p != NULL) {
-	dlist_insert_before(*d1p, d2);
-    }
-    *d1p = d2;
-}
-
-// pop the front of the dp list
-static inline DList * dlist_pop(DList **dp) {
-	//printf("%s(%p)\n", __FUNCTION__, dp);
-    DList *d1 = *dp;
-    DList *d2 = d1->next;
-    dlist_remove(d1);
-    if (d1 == d2) {
-	*dp = NULL;
-    } else {
-	*dp = d2;
-    }
-    return d1;
-}
-
-// remove d2 from the list, advancing d1p if needed
-static inline void dlist_remove_from(DList **d1p, DList *d2) {
-	//printf("%s(%p, %p)\n", __FUNCTION__, d1p, d2);
-    if (*d1p == d2) {
-	dlist_pop(d1p);
-    } else {
-	dlist_remove(d2);
-    }
-}
